Encapsulates Jwt system class
